Replaced calls to deprecated versions of createCell(), getCell(), createRow(), and getRow(). (Changing short to int)
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@688910 13f79535-47bb-0310-9956-ffa450edef68

+++ src/contrib/src/org/apache/poi/hssf/usermodel/contrib/HSSFCellUtil.java
@@ -113,7 +113,7 @@ public final class HSSFCellUtil
 
         if ( cell == null )
         {
-            cell = row.createCell( (short)column );
+            cell = row.createCell(column );
         }
         return cell;
     }
