+++ src/java/org/apache/poi/hssf/usermodel/HSSFSheet.java
@@ -1817,7 +1817,7 @@ public final class HSSFSheet {
         //  new HSSFComment instances, which is bad
         HSSFRow r = getRow(row);
         if(r != null) {
-            HSSFCell c = r.getCell((short)column);
+            HSSFCell c = r.getCell(column);
             if(c != null) {
                 return c.getCellComment();
             } else {
